Tidy namespaces
