+++ src/CollectionRenderer/ReferenceCollectionRenderer.php
@@ -1,15 +1,22 @@
 <?php
 
-namespace olvlvl\ComposerAttributeCollector;
+namespace olvlvl\ComposerAttributeCollector\CollectionRenderer;
 
-use function var_export;
+use olvlvl\ComposerAttributeCollector\CollectionRenderer;
+use olvlvl\ComposerAttributeCollector\TransientCollection;
+use olvlvl\ComposerAttributeCollector\TransientTargetClass;
+use olvlvl\ComposerAttributeCollector\TransientTargetMethod;
+use olvlvl\ComposerAttributeCollector\TransientTargetParameter;
+use olvlvl\ComposerAttributeCollector\TransientTargetProperty;
+
+use function in_array;
 
 /**
  * Renders collected attribute targets as PHP code.
  *
  * @internal
  */
-final class StaticCollectionRenderer
+final class ReferenceCollectionRenderer implements CollectionRenderer
 {
     public static function render(TransientCollection $collector): string
     {
@@ -24,7 +31,7 @@ final class StaticCollectionRenderer
 
         // attributes.php @generated by https://github.com/olvlvl/composer-attribute-collector
 
-        \olvlvl\ComposerAttributeCollector\Attributes::with(fn () => new \olvlvl\ComposerAttributeCollector\StaticCollection(
+        \olvlvl\ComposerAttributeCollector\Attributes::with(fn () => new \olvlvl\ComposerAttributeCollector\Collection\ReferenceCollection(
             targetClasses: $targetClassesCode,
             targetMethods: $targetMethodsCode,
             targetProperties: $targetPropertiesCode,
@@ -53,12 +60,10 @@ final class StaticCollectionRenderer
      * @param iterable<class-string, iterable<TransientTargetClass|TransientTargetMethod|TransientTargetParameter|TransientTargetProperty>> $targetByClass
      *
      * @return array<class-string, array<array{
-     *     array<int|string, mixed>,
      *     class-string,
-     *     2?:non-empty-string,
-     *     3?:non-empty-string
-     * }>> Where _key_ is an attribute class and _value_ is an array of parameters,
-     *     where `0` are the attribute arguments and `1` is the target class.
+     *     1?:non-empty-string,
+     *     2?:non-empty-string
+     * }>>
      */
     private static function targetsToArray(iterable $targetByClass): array
     {
@@ -66,9 +71,7 @@ final class StaticCollectionRenderer
 
         foreach ($targetByClass as $class => $targets) {
             foreach ($targets as $t) {
-                self::assertExportable($t->arguments);
-
-                $a = [ $t->arguments, $class ];
+                $a = [ $class ];
 
                 if ($t instanceof TransientTargetParameter) {
                     $a[] = $t->method;
@@ -82,25 +85,12 @@ final class StaticCollectionRenderer
                     $a[] = $t->name;
                 }
 
-                $by[$t->attribute][] = $a;
+                if (!in_array($a, $by[$t->attribute] ?? [], true)) {
+                    $by[$t->attribute][] = $a;
+                }
             }
         }
 
         return $by;
-    }
-
-    /**
-     * @param array<mixed> $arguments
-     */
-    private static function assertExportable(array $arguments): void
-    {
-        foreach ($arguments as $argument) {
-            if ($argument instanceof \Closure) {
-                throw new \InvalidArgumentException(
-                    "PHP 8.5 Closures in constant expressions are not supported by the 'static' strategy, use 'reference' instead." .
-                    " Please, check the README."
-                );
-            }
-        }
     }
 }
