+++ src/CollectionRenderer/StaticCollectionRenderer.php
@@ -1,15 +1,22 @@
 <?php
 
-namespace olvlvl\ComposerAttributeCollector;
+namespace olvlvl\ComposerAttributeCollector\CollectionRenderer;
 
-use function in_array;
+use olvlvl\ComposerAttributeCollector\CollectionRenderer;
+use olvlvl\ComposerAttributeCollector\TransientCollection;
+use olvlvl\ComposerAttributeCollector\TransientTargetClass;
+use olvlvl\ComposerAttributeCollector\TransientTargetMethod;
+use olvlvl\ComposerAttributeCollector\TransientTargetParameter;
+use olvlvl\ComposerAttributeCollector\TransientTargetProperty;
+
+use function var_export;
 
 /**
  * Renders collected attribute targets as PHP code.
  *
  * @internal
  */
-final class ReferenceCollectionRenderer
+final class StaticCollectionRenderer implements CollectionRenderer
 {
     public static function render(TransientCollection $collector): string
     {
@@ -24,7 +31,7 @@ final class ReferenceCollectionRenderer
 
         // attributes.php @generated by https://github.com/olvlvl/composer-attribute-collector
 
-        \olvlvl\ComposerAttributeCollector\Attributes::with(fn () => new \olvlvl\ComposerAttributeCollector\ReferenceCollection(
+        \olvlvl\ComposerAttributeCollector\Attributes::with(fn () => new \olvlvl\ComposerAttributeCollector\Collection\StaticCollection(
             targetClasses: $targetClassesCode,
             targetMethods: $targetMethodsCode,
             targetProperties: $targetPropertiesCode,
@@ -53,10 +60,12 @@ final class ReferenceCollectionRenderer
      * @param iterable<class-string, iterable<TransientTargetClass|TransientTargetMethod|TransientTargetParameter|TransientTargetProperty>> $targetByClass
      *
      * @return array<class-string, array<array{
+     *     array<int|string, mixed>,
      *     class-string,
-     *     1?:non-empty-string,
-     *     2?:non-empty-string
-     * }>>
+     *     2?:non-empty-string,
+     *     3?:non-empty-string
+     * }>> Where _key_ is an attribute class and _value_ is an array of parameters,
+     *     where `0` are the attribute arguments and `1` is the target class.
      */
     private static function targetsToArray(iterable $targetByClass): array
     {
@@ -64,7 +73,9 @@ final class ReferenceCollectionRenderer
 
         foreach ($targetByClass as $class => $targets) {
             foreach ($targets as $t) {
-                $a = [ $class ];
+                self::assertExportable($t->arguments);
+
+                $a = [ $t->arguments, $class ];
 
                 if ($t instanceof TransientTargetParameter) {
                     $a[] = $t->method;
@@ -78,12 +89,25 @@ final class ReferenceCollectionRenderer
                     $a[] = $t->name;
                 }
 
-                if (!in_array($a, $by[$t->attribute] ?? [], true)) {
-                    $by[$t->attribute][] = $a;
-                }
+                $by[$t->attribute][] = $a;
             }
         }
 
         return $by;
+    }
+
+    /**
+     * @param array<mixed> $arguments
+     */
+    private static function assertExportable(array $arguments): void
+    {
+        foreach ($arguments as $argument) {
+            if ($argument instanceof \Closure) {
+                throw new \InvalidArgumentException(
+                    "PHP 8.5 Closures in constant expressions are not supported by the 'static' strategy, use 'reference' instead." .
+                    " Please, check the README."
+                );
+            }
+        }
     }
 }
